requirements.txt
streamlit>=1.40.0
langchain-openai>=0.2.0
oci-openai>=0.1.0
httpx>=0.27.0
python-dotenv>=1.0.1
